move tracing.rs to execution mod
