+++ docs/source/tracing/paged.md
@@ -13,7 +13,7 @@ If tracing is enabled the row iterator will contain a list of tracing ids for al
 # use std::error::Error;
 # async fn check_only_compiles(session: &Session) -> Result<(), Box<dyn Error>> {
 use scylla::query::Query;
-use scylla::tracing::TracingInfo;
+use scylla::execution::tracing::TracingInfo;
 use futures::StreamExt;
 use uuid::Uuid;
 
@@ -51,7 +51,7 @@ for id in tracing_ids {
 # use std::error::Error;
 # async fn check_only_compiles(session: &Session) -> Result<(), Box<dyn Error>> {
 use scylla::prepared_statement::PreparedStatement;
-use scylla::tracing::TracingInfo;
+use scylla::execution::tracing::TracingInfo;
 use futures::StreamExt;
 use uuid::Uuid;
