+++ examples/tracing.rs
@@ -4,10 +4,10 @@
 use anyhow::{anyhow, Result};
 use futures::StreamExt as _;
 use scylla::batch::Batch;
+use scylla::execution::tracing::TracingInfo;
 use scylla::statement::{
     prepared_statement::PreparedStatement, query::Query, Consistency, SerialConsistency,
 };
-use scylla::tracing::TracingInfo;
 use scylla::QueryResult;
 use scylla::{Session, SessionBuilder};
 use std::env;
