+++ scylla/src/session/session_builder.rs
@@ -856,7 +856,7 @@ impl<K: SessionBuilderKind> GenericSessionBuilder<K> {
         self
     }
 
-    /// Set the number of attempts to fetch [TracingInfo](crate::tracing::TracingInfo)
+    /// Set the number of attempts to fetch [TracingInfo](crate::execution::tracing::TracingInfo)
     /// in [`Session::get_tracing_info`](crate::Session::get_tracing_info).
     /// The default is 5 attempts.
     ///
@@ -885,7 +885,7 @@ impl<K: SessionBuilderKind> GenericSessionBuilder<K> {
         self
     }
 
-    /// Set the delay between attempts to fetch [TracingInfo](crate::tracing::TracingInfo)
+    /// Set the delay between attempts to fetch [TracingInfo](crate::execution::tracing::TracingInfo)
     /// in [`Session::get_tracing_info`](crate::Session::get_tracing_info).
     /// The default is 3 milliseconds.
     ///
@@ -914,7 +914,7 @@ impl<K: SessionBuilderKind> GenericSessionBuilder<K> {
         self
     }
 
-    /// Set the consistency level of fetching [TracingInfo](crate::tracing::TracingInfo)
+    /// Set the consistency level of fetching [TracingInfo](crate::execution::tracing::TracingInfo)
     /// in [`Session::get_tracing_info`](crate::Session::get_tracing_info).
     /// The default is [`Consistency::One`].
     ///
